requirements.txt
pytesseract
Pillow
python-magic-bin
python-dateutil
supabase
python-dotenv
